+++ src/test/java/org/threadly/concurrent/wrapper/ThrowableSuppressingRunnableTest.java
@@ -1,13 +1,14 @@
-package org.threadly.concurrent;
+package org.threadly.concurrent.wrapper;
 
 import static org.junit.Assert.*;
 
 import org.junit.Test;
+import org.threadly.concurrent.TestRuntimeFailureRunnable;
 import org.threadly.test.concurrent.TestRunnable;
 import org.threadly.util.ExceptionUtils;
 import org.threadly.util.TestExceptionHandler;
 
-@SuppressWarnings({"javadoc", "deprecation"})
+@SuppressWarnings("javadoc")
 public class ThrowableSuppressingRunnableTest {
   @Test
   public void getContainedRunnableTest() {
